requirements.txt
fastapi
websockets
uvicorn
pyaudio
numpy==1.26.4
requests
pydub
openai
google-generativeai
mysql-connector-python
tiktoken
sounddevice
soundfile
silero-vad
onnxruntime
opencv-python==4.8.1.78
langchain
langchain-core
langgraph
langchain_google_genai
aiomysql